backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.0",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
